refactor: renamed interface parser file and moved it to generic folder
